# server/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Home API
  description: |
    API for controlling smart home devices including Philips Hue lights and Nanoleaf panels.
    
    ## Overview
    This API provides a unified interface for managing multiple smart lighting systems:
    - **Philips Hue** - Bridge-based smart lights with rooms/groups support
    - **Nanoleaf** - LED panels with effects support
    
    ## Authentication
    Device authentication is handled through pairing processes:
    - **Hue**: Press the link button on the bridge, then call `/api/hue/authenticate`
    - **Nanoleaf**: Hold power button for 5-7 seconds until lights flash, then call `/api/nanoleaf/authenticate`
    
    ## Color Modes
    Devices support multiple color modes:
    - `hs` - Hue/Saturation mode
    - `ct` - Color temperature mode (Kelvin)
    - `effect` - Device-specific effects
  version: 1.0.0
  contact:
    name: Smart Home API
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/devices/{id}:
    put:
      tags:
        - General
      summary: Update device state
      description: Update any device by ID (works for both Hue and Nanoleaf)
      operationId: updateDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceStateUpdate'
      responses:
        '200':
          description: Device updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Room Endpoints
  /api/rooms:
    get:
      tags:
        - Rooms
      summary: Get all rooms
      description: Returns all rooms with their devices and groups
      operationId: getRooms
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
    post:
      tags:
        - Rooms
      summary: Create a room
      description: Create a new room for organizing devices
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Room name
                  example: "Kitchen"
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
  /api/rooms/{id}:
    get:
      tags:
        - Rooms
      summary: Get a room
      description: Returns a specific room by ID
      operationId: getRoom
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Rooms
      summary: Update a room
      description: Update a room's name
      operationId: updateRoom
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: New room name
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Rooms
      summary: Delete a room
      description: Delete a room (devices will be unassigned)
      operationId: deleteRoom
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '204':
          description: Room deleted successfully
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/rooms/{roomId}/devices/{deviceExternalId}:
    put:
      tags:
        - Rooms
      summary: Assign device to room
      description: Assign a device to a room by its external ID
      operationId: assignDeviceToRoom
      parameters:
        - name: roomId
          in: path
          required: true
          description: Room ID
          schema:
            type: string
            format: uuid
        - name: deviceExternalId
          in: path
          required: true
          description: Device external ID (e.g., hue-1)
          schema:
            type: string
      responses:
        '200':
          description: Device assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Room or device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Rooms
      summary: Remove device from room
      description: Remove a device from a room (sets room to null)
      operationId: removeDeviceFromRoom
      parameters:
        - name: roomId
          in: path
          required: true
          description: Room ID
          schema:
            type: string
            format: uuid
        - name: deviceExternalId
          in: path
          required: true
          description: Device external ID (e.g., hue-1)
          schema:
            type: string
      responses:
        '200':
          description: Device removed from room successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Room or device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Device Group Endpoints
...
components:
  parameters:
    DeviceId:
      name: id
      in: path
      required: true
      description: Device ID (e.g., hue-1 for Hue lights or UUID for Nanoleaf)
      schema:
        type: string
        example: hue-1
...
    RoomId:
      name: id
      in: path
      required: true
      description: Room UUID
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440001
...
  schemas:
    Light:
      type: object
      description: Unified light/device representation
      properties:
        id:
          type: string
          description: Unique device identifier
          example: hue-1
        name:
          type: string
          description: User-friendly device name
          example: Living Room Lamp
        type:
          type: string
          enum: [hue, nanoleaf]
          description: Device type/manufacturer
        state:
          $ref: '#/components/schemas/DeviceState'
        reachable:
          type: boolean
          description: Whether the device is reachable on the network
        model:
          type: string
          description: Device model identifier
          example: LCT007
        capabilities:
          type: object
          properties:
            color:
              type: boolean
              description: Whether device supports color
            colorTemp:
              type: boolean
              description: Whether device supports color temperature
        roomId:
          type: string
          format: uuid
          nullable: true
          description: Room ID the device is assigned to
        roomName:
          type: string
          nullable: true
          description: Name of the room the device is assigned to
        groups:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
          description: Groups the device belongs to
    Room:
      type: object
      description: Room for organizing devices
      properties:
        id:
          type: string
          format: uuid
          description: Unique room identifier
        name:
          type: string
          description: Room name
          example: Kitchen
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Light'
          description: Devices in this room
        groups:
          type: array
          items:
            $ref: '#/components/schemas/DeviceGroup'
          description: Device groups in this room
    DeviceGroup:
      type: object
      description: Group of devices for simultaneous control
      properties:
        id:
          type: string
          format: uuid
          description: Unique group identifier
        name:
          type: string
          description: Group name
          example: Kitchen Table Lights
        roomId:
          type: string
          format: uuid
          nullable: true
          description: Room this group belongs to
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        room:
          $ref: '#/components/schemas/Room'
          nullable: true
          description: Room this group belongs to
        devices:
          type: array
          items:
            type: object
            properties:
              device:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  externalId:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
          description: Devices in this group
    DeviceState:
      type: object
      description: Current state of a device
      properties:
        on:
          type: boolean
          description: Whether the device is on
        brightness:
          type: integer
          minimum: 0
          maximum: 100
          description: Brightness level (0-100)
        color:
          $ref: '#/components/schemas/ColorHSB'
        colorTemp:
          type: integer
          minimum: 2000
          maximum: 6500
          description: Color temperature in Kelvin
          example: 4000
        colorMode:
          type: string
          enum: [hs, ct, effect]
          description: Current color mode
    DeviceStateUpdate:
      type: object
      description: Partial device state for updates
      properties:
        on:
          type: boolean
          description: Turn device on or off
        brightness:
          type: integer
          minimum: 0
          maximum: 100
          description: Brightness level (0-100)
        color:
          $ref: '#/components/schemas/ColorHSB'
        colorTemp:
          type: integer
          minimum: 2000
          maximum: 6500
          description: Color temperature in Kelvin
    ColorHSB:
      type: object
      description: HSB color representation
      properties:
        hue:
          type: number
          minimum: 0
          maximum: 360
          description: Hue value (0-360 degrees)
          example: 240
        saturation:
          type: number
          minimum: 0
          maximum: 100
          description: Saturation percentage (0-100)
          example: 100
        brightness:
          type: number
          minimum: 0
          maximum: 100
          description: Brightness percentage (0-100)
          example: 75
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Device not found